requirements.txt
3==0.0.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.1
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycodestyle==2.8.0
python-dotenv==0.20.0
requests==2.27.1
toml==0.10.2
urllib3==1.26.9
uWSGI==2.0.20
Werkzeug==2.1.0
zipp==3.7.0
